refactor(dal): restructure datatypes, contexts, and params
- Rename Blob to Object for consistency with object storage terminology
- Remove builder methods from all datatypes (Object, Document, Embedding, Record, Message, Graph, Node, Edge)
- Consolidate context files into core/contexts.rs with AnyContext enum
- Create core/params.rs with RelationalParams, ObjectParams, VectorParams
- Add bucket to ObjectParams, collection to VectorParams
- Move table from PostgresParams to RelationalParams
- Consolidate input_stream.rs and output_stream.rs into streams.rs
- Add Python generated types (datatypes.py, params.py, contexts.py)
- Use JsonValue type alias instead of Any for type safety
- Add PyO3 integration layer for Python provider implementations
- Fix tests to use struct literal initialization

+++ crates/nvisy-rig/src/rag/searcher/mod.rs
@@ -103,7 +103,7 @@ impl Searcher {
             .into_iter()
             .map(|scored| {
                 let chunk = scored.chunk;
-                let metadata = ChunkMetadata::from_json(&chunk.metadata, chunk.chunk_index);
+                let metadata = ChunkMetadata::from_json(&chunk.metadata, chunk.chunk_index as u32);
                 RetrievedChunk::new(chunk.id, chunk.file_id, scored.score, metadata)
             })
             .collect();
